perf(ci): shard the example-sweep gate across files for parallel runners
The physics-loop example-sweep gate ran `runValidateCli({ includeInterference:
true })` over every live example in ONE sequential test in ONE file. vitest
shards by file, so that file pinned a single CI runner and set the merge-gate
floor.

Extract discovery + classification + the per-example assertion into
exampleSweepShared.ts and split the heavy per-example checks across four
partition files (exampleSweepGate.shard{A,B,C,D}.test.ts), so vitest
distributes them across its per-file shards. The fast structural completeness
checks (no validate) stay in exampleSweepGate.test.ts.

Coverage is identical: the four partitions union to exactly the set the old
loop validated (verified — classification sets byte-identical to before;
modulo-N partition is a complete cover). Note: the live-validate set is
currently only two examples (so100 ~10s, spice-dispenser ~360s), so the present
floor is dominated by the single spice-dispenser interference sweep, which file
sharding cannot parallelize — that check's BREP cost (issue #348) is the
remaining lever. The split still distributes any future live examples and
separates the fast structural checks from the heavy validates.
